Fix SonarCloud findings (cycle 1)
